Refactor Game logic into managers
